fix: update phpstan version and remove redundant error from baseline
Updated the phpstan version in `.phive/phars.xml` from `1.10.67` to `1.12.13` to use the latest
features and improvements. Also, altered the specified version pattern to `1.*` to automatically
pick up the latest minor version releases. In addition, removed a redundant error message in
`phpstan-baseline.neon` related to the `getUrlRewrittenWithHost` method in `types/category.php`,
which was causing unnecessary noise in the error report.