fix: consider default currency and change event to price field
This commit adds the `getDefaultCurrency` function call and incorporates the default currency in
the price calculation. It further enhances the price input field to be disabled and auto-filled
when the default currency is in use. Changes to the primary input field are now also propagated to
the relevant currency input field.

In addition, the composer.json has been updated to require a minimum version of 2.1 for
`quiqqer/currency`.

Note: Besides, several function calls have been added to trigger `change` event on main input
element for accurate price calculation whenever the value changes. Moreover, the `blur` event has
been added to the input element in `onSubmit` event handler to reflect the changes immediately.