feat: add currency support for product calculation
This commit expands product calculations to include support for different currencies. The
'calc.php' file has been updated to now take a currency input and use it when calculating the
product cost. If no currency is given, it will default to the current user's currency.

The 'Product.js', 'Products.js' and 'frontend/Product.js' files have also been updated to pass the
currency when making Ajax calls.

Finally, if no currency is set for the product in 'ViewFrontend.php', the default currency or the
current user's currency will be used. Additionally, 'UniqueProduct.php' now includes the currency
code when converting to an article.

This feature enhances the flexibility of product cost calculations and accommodates for
international users by including their local currency in the calculation process.