fix: use runtime currency instead of default currency
This commit refractors how product currency is set by replacing the default currency with the
runtime currency which allows the application to adapt to different user sessions more responsively.

Additionally, a fallback mechanism for setting valueText in PriceFactor has been introduced in case
if it's given as a language-specific array.